Refactor repositories and services: update method signatures to use Long instead of String for Tid, rename Notification table and fields for clarity, and streamline database interactions in AlertService and JobContentService.

+++ src/main/java/com/wgzhao/addax/admin/model/SysDict.java
@@ -18,7 +18,7 @@ public class SysDict
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "code")
-    private Short code;
+    private Integer code;
 
     @Column(name = "name")
     private String name;
